Organized matrix lib files
